+++ crates/data/src/engine/mod.rs
@@ -719,6 +719,7 @@ impl DataEngine {
                         start_time_ns.into(),
                         None,
                         Some(callback),
+                        None,
                     )
                     .expect(FAILED);
             }